Improves the dumping process of the metaphysics schema to take stitching into account
