+++ _schema.graphql
@@ -1029,7 +1029,7 @@ type ArtworkContextFair {
     section: String
 
     # Sorts for shows in a fair
-    sort: ShowSort = FEATURED_DESC
+    sort: ShowSort = null
     after: String
     first: Int
     before: String
@@ -1958,7 +1958,7 @@ type Collection implements Node {
     before: String
     last: Int
     private: Boolean = false
-    sort: CollectionSorts = POSITION_DESC
+    sort: CollectionSorts = null
   ): ArtworkConnection
   description: String!
   default: Boolean!
@@ -2557,6 +2557,23 @@ type dimensions {
   cm: String
 }
 
+# Partner
+type DoNotUseThisPartner {
+  # Partners display name
+  display_name: String
+
+  # Partners given name
+  given_name: String
+
+  # Uniq ID for this partner
+  id: ID!
+  relative_size: Int
+
+  # Uniq, friendly identifier for this artist
+  slug: String!
+  subscription_state: String
+}
+
 type EcommerceError {
   # The error message
   type: String!
@@ -2670,7 +2687,7 @@ type Fair {
     section: String
 
     # Sorts for shows in a fair
-    sort: ShowSort = FEATURED_DESC
+    sort: ShowSort = null
     after: String
     first: Int
     before: String
@@ -3626,7 +3643,7 @@ type HomePageModuleContextFair {
     section: String
 
     # Sorts for shows in a fair
-    sort: ShowSort = FEATURED_DESC
+    sort: ShowSort = null
     after: String
     first: Int
     before: String
@@ -4294,6 +4311,16 @@ type MessageInitiator {
   email: String
 }
 
+type Metaphysics {
+  environment: Boolean!
+  queryTracing: Boolean!
+  headpDumps: Boolean!
+  stitching: Boolean!
+  stitchingGravity: Boolean!
+  stitchingConvection: Boolean!
+  stitchingExchange: Boolean!
+}
+
 type Mutation {
   # Create a bidder
   createBidder(input: CreateBidderInput!): CreateBidderPayload
@@ -4346,9 +4373,6 @@ type Mutation {
   createGeminiEntryForAsset(
     input: CreateGeminiEntryForAssetInput!
   ): CreateGeminiEntryForAssetPayload
-
-  # Records an artwork view.
-  recordArtworkView(input: RecordArtworkViewInput!): RecordArtworkViewPayload
 
   # Create a new consignment submission using Convection
   createConsignmentSubmission(
@@ -4414,6 +4438,7 @@ type Mutation {
 
   # Submits an order
   ecommerceSubmitOrder(input: SubmitOrderInput!): SubmitOrderPayload
+  recordArtworkView(input: RecordArtworkViewInput): RecordArtworkViewPayload
 }
 
 input Near {
@@ -5698,7 +5723,7 @@ type Query {
     sort: SaleSorts
   ): [Sale]
 
-  # The schema for difference microservice settings
+  # The schema for difference micro-service settings
   services: Services
 
   # A Show
@@ -5785,15 +5810,31 @@ type Query {
 
   # A wildcard used to support complex root queries in Relay
   viewer: Viewer
+
+  # Autocomplete resolvers.
+  _unused_gravity_match_partners(
+    term: String
+    page: Int = 1
+    size: Int = 5
+    match_type: String
+  ): [DoNotUseThisPartner]
+
+  # Find partners by ID
+  _unused_gravity_partners(ids: [ID]!): [DoNotUseThisPartner]
 }
 
+# Autogenerated input type of RecordArtworkView
 input RecordArtworkViewInput {
-  artwork_id: String!
+  # A unique identifier for the client performing the mutation.
   clientMutationId: String
+  artwork_id: String!
 }
 
+# Autogenerated return type of RecordArtworkView
 type RecordArtworkViewPayload {
   artwork_id: String!
+
+  # A unique identifier for the client performing the mutation.
   clientMutationId: String
 }
 
@@ -6306,6 +6347,7 @@ type SendConversationMessageMutationPayload {
 
 type Services {
   convection: Convection!
+  metaphysics: Metaphysics!
 }
 
 input SetOrderPaymentInput {
@@ -7296,7 +7338,7 @@ type Viewer {
     sort: SaleSorts
   ): [Sale]
 
-  # The schema for difference microservice settings
+  # The schema for difference micro-service settings
   services: Services
 
   # A Show
