+++ src/lib/stitching/mergeSchemas.ts
@@ -5,13 +5,13 @@ import { consignmentStitchingEnvironment } from "lib/stitching/convection/stitch
 import { executableExchangeSchema } from "lib/stitching/exchange/schema"
 import config from "config"
 
-import localSchema from "../../schema"
+import localSchema from "../../schema/schema"
 import { GraphQLSchema } from "graphql"
 
 /**
  * Incrementally merges in schemas according to `process.env`
  */
-export const incrementalMergeSchemas = async (testConfig?: any) => {
+export const incrementalMergeSchemas = (testConfig?: any) => {
   const environment = testConfig || config
 
   const {
@@ -25,18 +25,17 @@ export const incrementalMergeSchemas = async (testConfig?: any) => {
   const extensionResolvers = {} as any
 
   if (ENABLE_GRAVQL_ONLY_STITCHING) {
-    console.log("GRAV")
-    const gravitySchema = await executableGravitySchema()
+    const gravitySchema = executableGravitySchema()
     schemas.push(gravitySchema)
   }
 
   if (ENABLE_ECOMMERCE_STITCHING) {
-    const exchangeSchema = await executableExchangeSchema()
+    const exchangeSchema = executableExchangeSchema()
     schemas.push(exchangeSchema)
   }
 
   if (ENABLE_CONSIGNMENTS_STITCHING) {
-    const convectionSchema = await executableConvectionSchema()
+    const convectionSchema = executableConvectionSchema()
     schemas.push(convectionSchema)
 
     const { extensionSchema, resolvers } = consignmentStitchingEnvironment(
